+++ spec/integrations/shared/base_render_examples.rb
@@ -287,6 +287,36 @@ shared_examples 'Base::render' do
     end
   end
 
+  context 'Given blueprint has :root' do
+    let(:result) { '{"root":{"id":' + obj_id + ',"position_and_company":"Manager at Procore"}}' }
+    let(:blueprint) { blueprint_with_block }
+    it('returns json with a root') do
+      expect(blueprint.render(obj, root: :root)).to eq(result)
+    end
+  end
+
+  context 'Given blueprint has :meta' do
+    let(:result) { '{"root":{"id":' + obj_id + ',"position_and_company":"Manager at Procore"},"meta":"meta_value"}' }
+    let(:blueprint) { blueprint_with_block }
+    it('returns json with a root') do
+      expect(blueprint.render(obj, root: :root, meta: 'meta_value')).to eq(result)
+    end
+  end
+
+  context 'Given blueprint has :meta without :root' do
+    let(:blueprint) { blueprint_with_block }
+    it('raises a BlueprinterError') { 
+      expect{blueprint.render(obj, meta: 'meta_value')}.to raise_error(Blueprinter::BlueprinterError)
+    }
+  end
+
+  context 'Given blueprint has root as a non-supported object' do
+    let(:blueprint) { blueprint_with_block }
+    it('raises a BlueprinterError') { 
+      expect{blueprint.render(obj, root: {some_key: "invalid root"})}.to raise_error(Blueprinter::BlueprinterError)
+    }
+  end
+
   context 'Given blueprint has ::field with a block' do
     let(:result) { '{"id":' + obj_id + ',"position_and_company":"Manager at Procore"}' }
     let(:blueprint) { blueprint_with_block }